Refactor LLM provider and model utils into cli_utils.py
